Update: Added delete user route, update user route, learnt about security context.

+++ src/main/java/com/project/todo/jwtUtil/JwtUtil.java
@@ -19,7 +19,7 @@ public class JwtUtil {
     private String secret;
 
     @Value("${jwt.expiration}")
-    private int expiration;
+    private long expiration;
 
     // get key
     public Key getKey() {
